finally found out the slowness issue, the overlay was being allocated on every frame! grr

+++ example-OculusRiftRendering/src/testApp.h
@@ -49,5 +49,5 @@ public:
     ofVec3f cursorGaze;
 
     float hudZ;
-    float hudW, hudH;
+    int hudW, hudH;
 };
